refactor(LottoMachine): 패키지 변경
- 유틸 역할을 하는 클래스이기에 util 패키지로 이동
- utill 오타 수정
