+++ src/main/java/lotto/util/InputParser.java
@@ -1,4 +1,4 @@
-package lotto.utill;
+package lotto.util;
 
 import java.util.Arrays;
 import java.util.List;
